+++ CHANGELOG.md
@@ -5,6 +5,7 @@ FEATURES:
 * **Policies** Client methods for `policies` API endpoints
 * **Node Types** Client methods for `node_types` API endpoints
 * **Jobs** Client methods for `jobs` API endpoints
+* **Roles** Client methods for `roles` API endpoints
 
 ## 0.3 (May 17, 2021)